[FIX] website : fix visitor email and phone computing
When a lead entered a wrong email address in the contact form,
this crashes when loading visitor view.
We were doing a email concatenation to have all the email address
of the leads linked to the visitor.
But as it is a cornet case that shouls not happens often
(to have multiplus email address linked to the visitor)
we will only keep the last email addresse entered by the visitor.

Task ID: 2077182
PR #37472

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>