[FIX] stock, delivery: put in pack and discard
rev[0] writes the qty done on the move lines directly in put_in_pack,
meaning hitting discard in the delivery wizard left the move lines with
a changed qty_done. We fix that by editing the quantities in
_put_in_pack in the stock module and getting the right move lines in the
delivery modules: meaning, get the normal or suggested ones (a fix
missed by rev[1]) then with quantities or 0 quantities.

[0] c981c02505f6ab9e6ff8df43329064bf03413a86
[1] 4fde4af87991688d02f79cc9ecf1e421b81c7823

closes odoo/odoo#37563

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>