[FIX] web_editor: do not redraw option hierarchy on click
Before this commit, when a hierarchy of options was displayed in the
editor left panel and that the user click on any one, the related
snippet was enabled making all the children options disappear (like
the "go to parent" button in the old UI). Now we only keep this behavior
when we click on the title of the snippet's options.

closes odoo/odoo#37538

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>