[FIX] website_event: prevent crash if clicking on register too soon
Since https://github.com/odoo/odoo/commit/619d7a8eeb295008417ff14a8d9b92763d52534e,
the JS is now lazy loaded. As the "Register Now" button on each event
page relies on JavaScript to perform a RPC, it needs the JS to be loaded
to be able to work properly when clicking on it. In this case, clicking
on it while the JS is not loaded yet performs a standard form submit,
which lands on a route which cannot handle a standard form submit which
then results in a 400 error page.

This commit fixes the problem by adding the 'o_wait_lazy_js' class on
the button which simply prevents click on elements during lazy loading.

Discovered while working on task-2043872

closes odoo/odoo#37675

X-original-commit: https://github.com/odoo/odoo/commit/882ea5c513b12e179a5e7e74b95ce12c1860d1d3
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>