[FIX] base: debug assets in multi-website/multi-company
Before this commit, the assets were not found when the user is connected
in mode debug assets in a multi-website and multi-company. This occurs
because the attachments (assets) were created with the number of the
first website accessed, and cannot be loaded when trying to access the
second website.

The website_id is added in the creation of the views see:
c3db40ff0622040ad628d2b8c8d448ddec324329

Now, the attachments are created without the website number and is
shared between all the websites.

opw-2076768

closes odoo/odoo#37789

X-original-commit: 42eec88e5da164e4930c48113109f2c8d4b73da3
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>