[FIX] sale_stock: so line creation launch procurements twice
Use case to reproduce:
- Create a SO and confirm it
- Add a new line with a consumable/storable product

-> The delivery order contains the new lines but with the double of
required quantity.

Due to commit 099a728e3516d62da6dba791ada3a86cfe3f70b7 that try to lauch procurement in bacth.
Howerver The create from sale.order.line was not removed and the
_action_launch_procurement were launched twice (in sale.order write +
sale.order.line create).

The fix could remove the create function from sale.order.line but
the drawback is that everytime a sale.order.line is created/edited then
it should pass through the sale.order 'write' method.
That would create a lot of issue in the future compare to the small gain
of performance for a correction.

Task-2077129

closes odoo/odoo#38187

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>