[FIX] web: take care of "groups" attribute in search view arch
The attribute "groups" on filters and fields in a search view arch was not taken
into account anymore.
Now items that should be invisible are not anymore available for selection
in the interface (as expected) but are still activable as search defaults
(and the corresponding facets appear if needed).

Task ID: 2081450

closes odoo/odoo#42125

X-original-commit: d11dcecd72f8de6b1876b79f1ab53cd0a93b46a9
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>