[FIX] sale: Color column invisible in product attribute
Before this commit, when we click on a item in the tree view of
exclude item of product attribute value view, we got a traceback.
It come from the attribute column_invisible. This attribute was
used the remove the color column if the display type was not
color. But each line can have a display type different, so we
can't use invisible_column, we must use instead invisible
attribute.

After this commit, the invisible attribute is use, and each line
will display or not the color column, depends his display type.

closes odoo/odoo#38157

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>