[REM] web: remove unused controller /apps/<app>
This controller never really worked in the last 3 versions.
It was fixed in 11.3 with ba3ed61688d but broken with refactoring in 12.0 at a
higher level with d5fb61f1cd24.

It was even worse in 13.0 as it was leading to a traceback: `view_type` field
got removed with a5eb92da539fa but this controller was still reading that
field.

closes odoo/odoo#38355

X-original-commit: a5c4e262449fef03d05d5250e4523832db5fbf00
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>